Change data type of No. of units: Integer to Double. New functionality: Transfer of an investment (From one to another investor).

+++ src/main/java/org/sakuram/persmony/valueobject/InvestVO.java
@@ -22,7 +22,7 @@ public class InvestVO {
 	Boolean isAccrualApplicable;
 	Long bankDvId;
 	String investmentIdWithProvider;
-	int units;
+	double units;
 	Double faceValue, cleanPrice, accruedInterest, charges;
 	Double rateOfInterest;
 	Date investmentStartDate;
